% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InkAgro-package.R
\docType{package}
\name{InkAgro-package}
\alias{InkAgro}
\alias{InkAgro-package}
\title{InkAgro: análisis de experimentos agrícolas en español}
\description{
InkAgro analiza experimentos agrícolas con una sola llamada a
[inkagro()]: verifica que los datos correspondan al diseño, ajusta el
análisis de varianza, evalúa supuestos, señala valores atípicos y compara
medias, con un reporte en español pensado para quien no domina R ni
estadística.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bryanchavez03/InkAgro}
  \item Report bugs at \url{https://github.com/bryanchavez03/InkAgro/issues}
}

}
\author{
\strong{Maintainer}: Bryan Joseph Chavez Velaisosa \email{7366989122@untrm.edu.pe}

Authors:
\itemize{
  \item Bryan Joseph Chavez Velaisosa \email{7366989122@untrm.edu.pe}
}

}
\keyword{internal}
